Implement deamonizing, fixes #9 . Add logging. Add build command, fixes #4. Add notifications in settings, fixes #8 .
The script is now daemonizable via `./slight start`. Stop, restart and status has also been implemented.
If need be, one can run it with `./slight run` to run it non-daemonized.

The debug/print has been changed out with the logging module.

Notifications has been added in the settings, and are available via `get_notifications_settings`, but not
yet used in the system.

Finally, a build command has been added to the IRC bot, to enable building via IRC.